Remove the ZIPFoundation adapter
