+++ Support/Carthage/.xcodegen
@@ -139,10 +139,7 @@
           "framework" : "..\/..\/Carthage\/Build\/CryptoSwift.xcframework"
         },
         {
-          "framework" : "..\/..\/Carthage\/Build\/Fuzi.xcframework"
-        },
-        {
-          "framework" : "..\/..\/Carthage\/Build\/Minizip.xcframework"
+          "framework" : "..\/..\/Carthage\/Build\/ReadiumFuzi.xcframework"
         },
         {
           "framework" : "..\/..\/Carthage\/Build\/ReadiumZIPFoundation.xcframework"
@@ -173,10 +170,7 @@
           "framework" : "..\/..\/Carthage\/Build\/DifferenceKit.xcframework"
         },
         {
-          "framework" : "..\/..\/Carthage\/Build\/Fuzi.xcframework"
-        },
-        {
-          "framework" : "..\/..\/Carthage\/Build\/Minizip.xcframework"
+          "framework" : "..\/..\/Carthage\/Build\/ReadiumFuzi.xcframework"
         },
         {
           "framework" : "..\/..\/Carthage\/Build\/SwiftSoup.xcframework"
@@ -212,10 +206,7 @@
     "ReadiumOPDS" : {
       "dependencies" : [
         {
-          "framework" : "..\/..\/Carthage\/Build\/Fuzi.xcframework"
-        },
-        {
-          "framework" : "..\/..\/Carthage\/Build\/Minizip.xcframework"
+          "framework" : "..\/..\/Carthage\/Build\/ReadiumFuzi.xcframework"
         },
         {
           "target" : "ReadiumShared"
@@ -240,10 +231,10 @@
     "ReadiumShared" : {
       "dependencies" : [
         {
-          "framework" : "..\/..\/Carthage\/Build\/Fuzi.xcframework"
+          "framework" : "..\/..\/Carthage\/Build\/ReadiumFuzi.xcframework"
         },
         {
-          "framework" : "..\/..\/Carthage\/Build\/Minizip.xcframework"
+          "framework" : "..\/..\/Carthage\/Build\/ReadiumZIPFoundation.xcframework"
         },
         {
           "framework" : "..\/..\/Carthage\/Build\/SwiftSoup.xcframework"
@@ -274,10 +265,7 @@
           "framework" : "..\/..\/Carthage\/Build\/CryptoSwift.xcframework"
         },
         {
-          "framework" : "..\/..\/Carthage\/Build\/Fuzi.xcframework"
-        },
-        {
-          "framework" : "..\/..\/Carthage\/Build\/Minizip.xcframework"
+          "framework" : "..\/..\/Carthage\/Build\/ReadiumFuzi.xcframework"
         },
         {
           "target" : "ReadiumShared"
@@ -13974,7 +13962,6 @@
 ../../Sources/Shared/Toolkit/XML/Fuzi.swift
 ../../Sources/Shared/Toolkit/XML/XML.swift
 ../../Sources/Shared/Toolkit/ZIP
-../../Sources/Shared/Toolkit/ZIP/Minizip.swift
 ../../Sources/Shared/Toolkit/ZIP/ZIPArchiveOpener.swift
 ../../Sources/Streamer
 ../../Sources/Streamer/Assets
