+++ Support/Carthage/project.yml
@@ -14,6 +14,7 @@ targets:
       - path: ../../Sources/Shared
     dependencies:
       - framework: ../../Carthage/Build/ReadiumFuzi.xcframework
+      - framework: ../../Carthage/Build/ReadiumZIPFoundation.xcframework
       - framework: ../../Carthage/Build/SwiftSoup.xcframework
       - target: ReadiumInternal
       - sdk: CoreServices.framework
