+++ TestApp/Integrations/Local/project+lcp.yml
@@ -52,8 +52,6 @@ targets:
       - package: Readium
         product: ReadiumAdapterLCPSQLite
       - package: Readium
-        product: ReadiumAdapterZIPFoundation
-      - package: Readium
         product: ReadiumOPDS
       - package: Readium
         product: ReadiumLCP
